comments for unit tests are all up. moved everything to a folder for readability
