--- Chupacabra/src/data/DataServiceTest.java
@@ -1,25 +1,0 @@
-//not totally sure what I should be testing here.
-package data;
-
-import static org.junit.Assert.*;
-
-import org.junit.Before;
-import org.junit.Test;
-
-public class DataServiceTest {
-
-	@Before
-	public void setUp() throws Exception {
-	}
-
-	@Test
-	public void testLoadDBs() {
-		fail("Not yet implemented");
-	}
-
-	@Test
-	public void testBackup() {
-		fail("Not yet implemented");
-	}
-
-}
